Split `cabal-validate` into modules
This disentangles the utility boilerplate from the validation logic,
making the `Main.hs` module much easier to read and modify.
