+++ cabal-validate/src/Cli.hs
@@ -1,0 +1,381 @@
+module Cli
+  ( Opts (..)
+  , parseOpts
+  , HackageTests (..)
+  , Compiler (..)
+  , VersionParseException (..)
+  )
+where
+
+import Control.Applicative (Alternative (many, (<|>)), (<**>))
+import Control.Exception (Exception (displayException), throw)
+import Control.Monad (forM_, when)
+import Data.Data (Typeable)
+import Data.Maybe (listToMaybe)
+import qualified Data.Text as T
+import qualified Data.Text.Lazy as T (toStrict)
+import qualified Data.Text.Lazy.Encoding as T (decodeUtf8)
+import Data.Version (Version, parseVersion)
+import GHC.Conc (getNumCapabilities)
+import Options.Applicative
+  ( FlagFields
+  , Mod
+  , Parser
+  , ParserInfo
+  , auto
+  , execParser
+  , flag
+  , flag'
+  , fullDesc
+  , help
+  , helper
+  , hidden
+  , info
+  , long
+  , maybeReader
+  , option
+  , progDesc
+  , short
+  , strOption
+  , switch
+  , value
+  )
+import qualified Options.Applicative as Opt
+import System.Directory (getCurrentDirectory)
+import System.Exit (exitSuccess)
+import System.Info (arch, os)
+import System.Process.Typed (proc, readProcessStdout_)
+import Text.ParserCombinators.ReadP (readP_to_S)
+
+import ClockUtil (AbsoluteTime, getAbsoluteTime)
+import Step (Step (..), displayStep, parseStep)
+
+-- | Command-line options, resolved with context from the environment.
+data Opts = Opts
+  { verbose :: Bool
+  , jobs :: Int
+  , cwd :: FilePath
+  , startTime :: AbsoluteTime
+  , compiler :: Compiler
+  , extraCompilers :: [FilePath]
+  , cabal :: FilePath
+  , hackageTests :: HackageTests
+  , archPath :: FilePath
+  , projectFile :: FilePath
+  , tastyArgs :: [String]
+  , targets :: [String]
+  , steps :: [Step]
+  }
+  deriving (Show)
+
+data HackageTests
+  = CompleteHackageTests
+  | PartialHackageTests
+  | NoHackageTests
+  deriving (Show)
+
+data Compiler = Compiler
+  { compilerExecutable :: FilePath
+  , compilerVersion :: Version
+  }
+  deriving (Show)
+
+data VersionParseException = VersionParseException
+  { versionInput :: String
+  , versionExecutable :: FilePath
+  }
+  deriving (Typeable, Show)
+
+instance Exception VersionParseException where
+  displayException exception =
+    "Failed to parse `"
+      <> versionExecutable exception
+      <> " --numeric-version` output: "
+      <> show (versionInput exception)
+
+makeCompiler :: FilePath -> IO Compiler
+makeCompiler executable = do
+  stdout <-
+    readProcessStdout_ $
+      proc executable ["--numeric-version"]
+  let version = T.unpack $ T.strip $ T.toStrict $ T.decodeUtf8 stdout
+      parsedVersions = readP_to_S parseVersion version
+      -- Who needs error messages? Those aren't in the API.
+      maybeParsedVersion =
+        listToMaybe
+          [ parsed
+          | (parsed, []) <- parsedVersions
+          ]
+      parsedVersion = case maybeParsedVersion of
+        Just parsedVersion' -> parsedVersion'
+        Nothing ->
+          throw
+            VersionParseException
+              { versionInput = version
+              , versionExecutable = executable
+              }
+
+  pure
+    Compiler
+      { compilerExecutable = executable
+      , compilerVersion = parsedVersion
+      }
+
+resolveOpts :: RawOpts -> IO Opts
+resolveOpts opts = do
+  let optionals :: Bool -> [a] -> [a]
+      optionals True items = items
+      optionals False _ = []
+
+      optional :: Bool -> a -> [a]
+      optional keep item = optionals keep [item]
+
+      steps' =
+        if not (null (rawSteps opts))
+          then rawSteps opts
+          else
+            concat
+              [
+                [ PrintConfig
+                , PrintToolVersions
+                , Build
+                ]
+              , optional (rawDoctest opts) Doctest
+              , optional (rawRunLibTests opts) LibTests
+              , optional (rawRunLibSuite opts) LibSuite
+              , optional (rawRunLibSuite opts && not (null (rawExtraCompilers opts))) LibSuiteExtras
+              , optional (rawRunCliTests opts && not (rawLibOnly opts)) CliTests
+              , optional (rawRunCliSuite opts && not (rawLibOnly opts)) CliSuite
+              , optionals (rawSolverBenchmarks opts) [SolverBenchmarksTests, SolverBenchmarksRun]
+              , [TimeSummary]
+              ]
+
+      targets' =
+        concat
+          [
+            [ "Cabal"
+            , "Cabal-hooks"
+            , "cabal-testsuite"
+            , "Cabal-tests"
+            , "Cabal-QuickCheck"
+            , "Cabal-tree-diff"
+            , "Cabal-described"
+            ]
+          , optionals
+              (CliTests `elem` steps')
+              [ "cabal-install"
+              , "cabal-install-solver"
+              , "cabal-benchmarks"
+              ]
+          , optional (rawSolverBenchmarks opts) "solver-benchmarks"
+          ]
+
+      archPath' =
+        let osPath =
+              case os of
+                "darwin" -> "osx"
+                "linux" -> "linux"
+                "mingw32" -> "windows"
+                _ -> os -- TODO: Warning?
+         in arch <> "-" <> osPath
+
+      projectFile' =
+        if rawLibOnly opts
+          then "cabal.validate-libonly.project"
+          else "cabal.validate.project"
+
+      tastyArgs' =
+        "--hide-successes"
+          : case rawTastyPattern opts of
+            Just tastyPattern -> ["--pattern", tastyPattern]
+            Nothing -> []
+
+  when (rawListSteps opts) $ do
+    -- TODO: This should probably list _all_ available steps, not just the selected ones!
+    putStrLn "Targets:"
+    forM_ targets' $ \target -> do
+      putStrLn $ "  " <> target
+    putStrLn "Steps:"
+    forM_ steps' $ \step -> do
+      putStrLn $ "  " <> displayStep step
+    exitSuccess
+
+  startTime' <- getAbsoluteTime
+  jobs' <- maybe getNumCapabilities pure (rawJobs opts)
+  cwd' <- getCurrentDirectory
+  compiler' <- makeCompiler (rawCompiler opts)
+
+  pure
+    Opts
+      { verbose = rawVerbose opts
+      , jobs = jobs'
+      , cwd = cwd'
+      , startTime = startTime'
+      , compiler = compiler'
+      , extraCompilers = rawExtraCompilers opts
+      , cabal = rawCabal opts
+      , archPath = archPath'
+      , projectFile = projectFile'
+      , hackageTests = rawHackageTests opts
+      , tastyArgs = tastyArgs'
+      , targets = targets'
+      , steps = steps'
+      }
+
+-- | Command-line options.
+data RawOpts = RawOpts
+  { rawVerbose :: Bool
+  , rawJobs :: Maybe Int
+  , rawCompiler :: FilePath
+  , rawCabal :: FilePath
+  , rawExtraCompilers :: [FilePath]
+  , rawTastyPattern :: Maybe String
+  , rawDoctest :: Bool
+  , rawSteps :: [Step]
+  , rawListSteps :: Bool
+  , rawLibOnly :: Bool
+  , rawRunLibTests :: Bool
+  , rawRunCliTests :: Bool
+  , rawRunLibSuite :: Bool
+  , rawRunCliSuite :: Bool
+  , rawSolverBenchmarks :: Bool
+  , rawHackageTests :: HackageTests
+  }
+  deriving (Show)
+
+optsParser :: Parser RawOpts
+optsParser =
+  RawOpts
+    <$> ( flag'
+            True
+            ( short 'v'
+                <> long "verbose"
+                <> help "Always display build and test output"
+            )
+            <|> flag
+              False
+              False
+              ( short 'q'
+                  <> long "quiet"
+                  <> help "Silence build and test output"
+              )
+        )
+    <*> option
+      (Just <$> auto)
+      ( short 'j'
+          <> long "jobs"
+          <> help "Passed to `cabal build --jobs`"
+          <> value Nothing
+      )
+    <*> strOption
+      ( short 'w'
+          <> long "with-compiler"
+          <> help "Build Cabal with the given compiler instead of `ghc`"
+          <> value "ghc"
+      )
+    <*> strOption
+      ( long "with-cabal"
+          <> help "Test the given `cabal-install` (the `cabal` on your `$PATH` is used for builds)"
+          <> value "cabal"
+      )
+    <*> many
+      ( strOption
+          ( long "extra-hc"
+              <> help "Extra compilers to run the test suites against"
+          )
+      )
+    <*> option
+      (Just <$> Opt.str)
+      ( short 'p'
+          <> long "pattern"
+          <> help "Pattern to filter tests by"
+          <> value Nothing
+      )
+    <*> boolOption
+      False
+      "doctest"
+      ( help "Run doctest on the `Cabal` library"
+      )
+    <*> many
+      ( option
+          (maybeReader parseStep)
+          ( short 's'
+              <> long "step"
+              <> help "Run only a specific step (can be specified multiple times)"
+          )
+      )
+    <*> switch
+      ( long "list-steps"
+          <> help "List the available steps and exit"
+      )
+    <*> ( flag'
+            True
+            ( long "lib-only"
+                <> help "Test only `Cabal` (the library)"
+            )
+            <|> flag
+              False
+              False
+              ( long "cli"
+                  <> help "Test `cabal-install` (the executable) in addition to `Cabal` (the library)"
+              )
+        )
+    <*> boolOption
+      True
+      "run-lib-tests"
+      ( help "Run tests for the `Cabal` library"
+      )
+    <*> boolOption
+      True
+      "run-cli-tests"
+      ( help "Run client tests for the `cabal-install` executable"
+      )
+    <*> boolOption
+      False
+      "run-lib-suite"
+      ( help "Run `cabal-testsuite` with the `Cabal` library"
+      )
+    <*> boolOption
+      False
+      "run-cli-suite"
+      ( help "Run `cabal-testsuite` with the `cabal-install` executable"
+      )
+    <*> boolOption
+      False
+      "solver-benchmarks"
+      ( help "Build and trial run `solver-benchmarks`"
+      )
+    <*> ( flag'
+            CompleteHackageTests
+            ( long "complete-hackage-tests"
+                <> help "Run `hackage-tests` on complete Hackage data"
+            )
+            <|> flag
+              NoHackageTests
+              PartialHackageTests
+              ( long "partial-hackage-tests"
+                  <> help "Run `hackage-tests` on parts of Hackage data"
+              )
+        )
+
+-- | Parse a boolean switch with separate names for the true and false options.
+boolOption' :: Bool -> String -> String -> Mod FlagFields Bool -> Parser Bool
+boolOption' defaultValue trueName falseName modifiers =
+  flag' True (modifiers <> long trueName)
+    <|> flag defaultValue False (modifiers <> hidden <> long falseName)
+
+-- | Parse a boolean switch with a `--no-*` flag for setting the option to false.
+boolOption :: Bool -> String -> Mod FlagFields Bool -> Parser Bool
+boolOption defaultValue trueName =
+  boolOption' defaultValue trueName ("no-" <> trueName)
+
+fullOptsParser :: ParserInfo RawOpts
+fullOptsParser =
+  info
+    (optsParser <**> helper)
+    ( fullDesc
+        <> progDesc "Test suite runner for `Cabal` and `cabal-install` developers"
+    )
+
+parseOpts :: IO Opts
+parseOpts = execParser fullOptsParser >>= resolveOpts
